Fix AES encryption streaming support
The update_aes_extra_data function requires the underlying writer to
support streaming to succeed. In effect, this function is only used to
switch the mode between AE2 and AE1 depending on file size. This is not
strictly necessary so long as we use the strong mode by default, so we
now only call this function for the non-streaming case.